reorganiza controllers e implementa bootstrap
