+++ controller/__init__.py
@@ -1,0 +1,2 @@
+from . import product
+from . import user
