+++ url/url.py
@@ -1,14 +1,14 @@
-from controller import controller
+from controller import product, user
 
 urls = [
-    ('/', controller.index),
-    ('/index', controller.index),
-    ('/signup', controller.signup),
-    ('/signin', controller.signin),
-    ('/products', controller.products),
-    ('/products/new', controller.new_product),
-    ('/products/<key>', controller.get_product_by_id),
-    ('/products/<key>/delete', controller.delete_product),
+    ('/', user.index),
+    ('/index', user.index),
+    ('/signup', user.signup),
+    ('/signin', user.signin),
+    ('/products', product.products),
+    ('/products/new', product.new_product),
+    ('/products/<key>', product.get_product_by_id),
+    ('/products/<key>/delete', product.delete_product),
 ]
