+++ view/templates/product/detail.html
@@ -2,20 +2,24 @@
 {% block title %}{{ product.name }}{% endblock %}
 
 {% block content %}
-    <div class="container">
-        <h1>Details</h1>
-        <div class="product-details">
-            <p><strong>Product ID:</strong> {{ product.product_id }}</p>
-            <p><strong>Name:</strong> {{ product.name }}</p>
-            <p><strong>Description:</strong> {{ product.description }}</p>
-            <p><strong>Price:</strong> ${{ product.price }}</p>
-            <p><strong>Stock Quantity:</strong> {{ product.stock_quantity }}</p>
-        </div>
-        <a href="/products">Back to Product List</a>
-        <form action="/products/{{ product.product_id }}/delete" method="POST">
-            <button type="submit" onclick="return confirm('Are you sure you want to delete this product?');">Delete Product</button>
-        </form>
-
+    <h1>Details</h1>
+    <div class="product-details">
+        <p><strong>Product ID:</strong> {{ product.product_id }}</p>
+        <p><strong>Name:</strong> {{ product.name }}</p>
+        <p><strong>Description:</strong> {{ product.description }}</p>
+        <p><strong>Price:</strong> ${{ product.price }}</p>
+        <p><strong>Stock Quantity:</strong> {{ product.stock_quantity }}</p>
     </div>
+    <div class="d-flex flex-row">
+        <div class="me-2">
+            <a href="/products" class="btn btn-secondary">Back to Product List</a>
+        </div>
+        <div>
+            <form action="/products/{{ product.product_id }}/delete" method="POST">
+                <button type="submit" onclick="return confirm('Are you sure you want to delete this product?');" class="btn btn-danger">Delete Product</button>
+            </form>
+        </div>
+    </div>
+
 {% endblock %}
