+++ view/templates/product/new.html
@@ -2,24 +2,22 @@
 {% block title %}nuevo producto{% endblock %}
 
 {% block content %}
-    <h1>Nuevo producto</h1>
-    <form action="/products" method="post">
-        <label for="name">Nombre:</label>
-        <input type="text" id="name" name="name" required autofocus>
-        <br><br>
+    <div class="w-25">
+        <h1>Nuevo producto</h1>
+        <form action="/products" method="post">
+            <label for="name" class="form-label">Nombre:</label>
+            <input type="text" id="name" name="name" class="form-control" required autofocus>
 
-        <label for="description">Descripcion:</label>
-        <input type="text" id="description" name="description" required>
-        <br><br>
+            <label for="description" class="form-label">Descripcion:</label>
+            <input type="text" id="description" name="description" class="form-control" required>
 
-        <label for="price">Precio:</label>
-        <input type="text" id="price" name="price" required>
-        <br><br>
+            <label for="price" class="form-label">Precio:</label>
+            <input type="text" id="price" name="price" class="form-control" required>
 
-        <label for="stock_quantity">Cantidad:</label>
-        <input type="text" id="stock_quantity" name="stock_quantity" required>
-        <br><br>
+            <label for="stock_quantity" class="form-label">Cantidad:</label>
+            <input type="text" id="stock_quantity" name="stock_quantity" class="form-control" required>
 
-        <input type="submit" value="Agregar">
-    </form>
+            <input type="submit" value="Agregar" class="mt-2 btn btn-primary">
+        </form>
+    </div>
 {% endblock %}
