+++ view/templates/product/search.html
@@ -2,5 +2,5 @@
       hx-target="#products-table">
     <label for="name">Nombre:</label>
     <input type="text" id="name" name="name" required>
-    <input type="submit" value="Buscar">
+    <input type="submit" value="Buscar" class="btn btn-secondary">
 </form>
